requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
folium>=0.14.0
geopy>=2.3.0
pydub>=0.25.1
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
